refactor: 클래스명 컨벤션에 맞게 level usecase 수정
