--- usecase/src/main/java/org/depromeet/spot/usecase/port/in/member/LevelUsecase.java
@@ -1,11 +1,0 @@
-package org.depromeet.spot.usecase.port.in.member;
-
-import java.util.List;
-
-import org.depromeet.spot.domain.member.Level;
-
-public interface LevelUsecase {
-    List<Level> findAllLevels();
-
-    Level findLevelUpDialogInfo(int nextLevel);
-}
