--- usecase/src/main/java/org/depromeet/spot/usecase/port/in/member/ReadLevelUsecase.java
@@ -1,10 +1,0 @@
-package org.depromeet.spot.usecase.port.in.member;
-
-import org.depromeet.spot.domain.member.Level;
-
-public interface ReadLevelUsecase {
-
-    Level findInitialLevel();
-
-    Level findByValue(int value);
-}
