--- usecase/src/main/java/org/depromeet/spot/usecase/service/level/LevelService.java
@@ -1,33 +1,0 @@
-package org.depromeet.spot.usecase.service.level;
-
-import java.util.List;
-
-import org.depromeet.spot.domain.member.Level;
-import org.depromeet.spot.usecase.port.in.member.LevelUsecase;
-import org.depromeet.spot.usecase.port.in.member.ReadLevelUsecase;
-import org.depromeet.spot.usecase.port.out.member.LevelRepository;
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class LevelService implements LevelUsecase {
-
-    private final LevelRepository levelRepository;
-    private final ReadLevelUsecase readLevelUsecase;
-
-    @Override
-    @Cacheable(cacheNames = {"levelsCache"})
-    public List<Level> findAllLevels() {
-        return levelRepository.findAll();
-    }
-
-    @Override
-    public Level findLevelUpDialogInfo(final int nextLevel) {
-        return readLevelUsecase.findByValue(nextLevel);
-    }
-}
